9.September_ai_doc_summariser/requirements.txt
Flask==3.1.1
Flask-WTF==1.2.2
whisper==1.1.10
Werkzeug==3.1.3
google-auth==2.38.0
google-genai==1.24.0
WTForms==3.2.1
SpeechRecognition==3.14.3
PyAudio==0.2.14
pydub==0.25.1